pitch-deck-debater/requirements.txt
streamlit>=1.28.0
anthropic>=0.34.0
python-pptx>=0.6.21
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.17.0
edge-tts
Pillow
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
